Add checkAuth utility
checkAuth is a module that contains a function to get token from local
storage, a function to send request to server for verification and
a function to remove token from local storage if it is invalid

Function checkAuth returns true of false whether token is valid or not.
Returned value is used to set authentication context in AuthContext.